src/requirements.txt
python-dotenv
aiohttp
teams-ai>=1.6.0,<2.0.0
requests
urllib3
flask